Update and rename NPC_light_ver.py to NPC_simple_ver.py
I combined all models in only one model
